requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyAutoGUI~=0.9.53
keyboard~=0.13.5